+++ src/main/java/mlplanets/service/CelestialObjectService.java
@@ -6,6 +6,6 @@ import java.awt.geom.Point2D;
 
 public interface CelestialObjectService {
 
-    Point2D getPositionForDay(CelestialObject object, int day);
+    Point2D getPositionForDay(CelestialObject object, long day);
 
 }
