+++ src/main/java/mlplanets/service/OrbitService.java
@@ -6,8 +6,8 @@ import java.awt.geom.Point2D;
 
 public interface OrbitService {
 
-    Point2D getPositionForDay(Orbit orbit, int day);
+    Point2D getPositionForDay(Orbit orbit, long day);
 
-    int getYearDurationInDays(Orbit orbit);
+    long getYearDurationInDays(Orbit orbit);
 
 }
